requirements.txt
# Core
streamlit>=1.31.0
langgraph>=0.0.48
openai>=1.14.0
tavily-python>=0.3.3
pydantic>=2.6.4
python-dotenv>=1.0.1

# Report generation
reportlab>=4.1.0
markdown>=3.5.2

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
